timetable_solutions/test-requirements.txt
# Requirements for running the test suite
django-webtest==1.9.10
factory_boy==3.2.1
pytest==7.1.3
pytest-cov==4.0.0
pytest-django==4.5.2
